include collection promotion
